// frontend/src/api/types.ts
export enum Role {
  SUPER_ADMIN = "super_admin",
  ADMIN = "admin",
  OPERATOR = "operator",
  VIEWER = "viewer",
  QA_MANAGER = "qa_manager",
}
// ...
export enum AIProcessingMode {
  AUTOMATIC = "automatic",
  MANUAL = "manual",
  NOT_ALLOWED = "not_allowed",
}